Fix a buffer overflow
hid-libusb.c:223: warning: array subscript is above array bounds

sizeof(wbuf) gives the size of the wbuf array in char. It is also the
number of element only if wbuf is an array of char.

Here wbuf is an array of wchar_t so the last element of the array is
wbuf[sizeof(wbuf)/sizeof(wbuf[0])-1]

Signed-off-by: Alan Ott <alan@signal11.us>