Do not use printf() unconditionally
The problem is that the file descriptor 1 used by printf() may not be
stdout but somethink else. For example a daemon may close fds 0, 1 and 2
and reuse them to open files or sockets. So using printf() may send data
on a file or socket opened by the application.

Errors will now be logged to stderr only if DEBUG_PRINTF is defined.