mac: Fix incorrect device list after device removal
On Mac OS X 10.8.x (Mountain Lion), if a device is connected, opened, then
closed, then unplugged, subsequent calls to hid_enumerate() will still
list the device (even though it's been unplugged).  If the device is
plugged in again, a second instance will show up in the list and often it
will be impossible to open either.  Github user TamToucan figured out that
in hid_enumerate() if a call is made to IOHIDManagerSetDeviceMatching()
before the call to IOHIDManagerCopyDevices() that this will clear the
problem.

Thanks to TamToucan for figuring this out.